# docs/swagger.yaml
basePath: /
definitions:
  dangbamgong-backend_internal_domain.ErrorCode:
    enum:
    - BAD_REQUEST
    - NOT_FOUND
    - INTERNAL_SERVER_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - CONFLICT
    - SERVICE_UNAVAILABLE
    - INVALID_TOKEN
    - INVALID_NICKNAME
    - NICKNAME_ALREADY_SET
    - USER_NOT_FOUND
    - ALREADY_BLOCKED
    - NOT_BLOCKED
    - ALREADY_IN_VOID
    - NOT_IN_VOID
    - TOO_MANY_ACTIVITIES
    - INVALID_ACTIVITY_NAME
    - ACTIVITY_ALREADY_EXISTS
    - ACTIVITY_NOT_FOUND
    - NOTIFICATION_NOT_FOUND
    - ALREADY_FRIENDS
    - REQUEST_ALREADY_SENT
    - BLOCKED
    - REQUEST_NOT_FOUND
    - REQUEST_NOT_PENDING
    - NOT_FRIENDS
    - FRIEND_NOT_IN_VOID
    - INVALID_REQUEST_TYPE
    type: string
    x-enum-varnames:
    - ErrBadRequest
    - ErrNotFound
    - ErrInternalServer
    - ErrUnauthorized
    - ErrForbidden
    - ErrConflict
    - ErrServiceUnavailable
    - ErrInvalidToken
    - ErrInvalidNickname
    - ErrNicknameAlreadySet
    - ErrUserNotFound
    - ErrAlreadyBlocked
    - ErrNotBlocked
    - ErrAlreadyInVoid
    - ErrNotInVoid
    - ErrTooManyActivities
    - ErrInvalidActivityName
    - ErrActivityAlreadyExists
    - ErrActivityNotFound
    - ErrNotificationNotFound
    - ErrAlreadyFriends
    - ErrRequestAlreadySent
    - ErrBlocked
    - ErrRequestNotFound
    - ErrRequestNotPending
    - ErrNotFriends
    - ErrFriendNotInVoid
    - ErrInvalidRequestType
...
  dangbamgong-backend_internal_dto.BlockItem:
    properties:
      blockedAt:
        type: string
      nickname:
        type: string
      tag:
        type: string
      userId:
        type: string
    type: object
  dangbamgong-backend_internal_dto.BlockListResponse:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.BlockItem'
        type: array
    type: object
...
  dangbamgong-backend_internal_dto.ChangeNicknameRequest:
    properties:
      nickname:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - nickname
    type: object
  dangbamgong-backend_internal_dto.ChangeNicknameResponse:
    properties:
      nickname:
        type: string
    type: object
...
  dangbamgong-backend_internal_dto.ErrorResponse:
    properties:
      code:
        $ref: '#/definitions/dangbamgong-backend_internal_domain.ErrorCode'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.NotificationSettings:
    properties:
      friendNudge:
        type: boolean
      friendRequest:
        type: boolean
      reminderHours:
        type: integer
      voidReminder:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-any:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_BlockListResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.BlockListResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_ChangeNicknameResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.ChangeNicknameResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UpdateSettingsResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.UpdateSettingsResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UserMeResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.UserMeResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UserSearchResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.UserSearchResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidEndResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.VoidEndResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidHistoryResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.VoidHistoryResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidStartResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.VoidStartResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.TestVoidRequest:
    properties:
      activities:
        items:
          type: string
        type: array
      endedAt:
        type: string
      startedAt:
        type: string
    type: object
...
  dangbamgong-backend_internal_dto.UpdateSettingsRequest:
    properties:
      friendNudge:
        type: boolean
      friendRequest:
        type: boolean
      reminderHours:
        type: integer
      voidReminder:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.UpdateSettingsResponse:
    properties:
      friendNudge:
        type: boolean
      friendRequest:
        type: boolean
      reminderHours:
        type: integer
      voidReminder:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.UserMeResponse:
    properties:
      currentVoidStartedAt:
        type: string
      id:
        type: string
      isInVoid:
        type: boolean
      nickname:
        type: string
      notificationSettings:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.NotificationSettings'
      tag:
        type: string
    type: object
  dangbamgong-backend_internal_dto.UserSearchItem:
    properties:
      isBlocked:
        type: boolean
      nickname:
        type: string
      tag:
        type: string
      userId:
        type: string
    type: object
  dangbamgong-backend_internal_dto.UserSearchResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.UserSearchItem'
        type: array
    type: object
  dangbamgong-backend_internal_dto.VoidEndRequest:
    properties:
      activities:
        items:
          type: string
        maxItems: 5
        type: array
    type: object
  dangbamgong-backend_internal_dto.VoidEndResponse:
    properties:
      activities:
        items:
          type: string
        type: array
      durationSec:
        type: integer
      endedAt:
        type: string
      sessionId:
        type: string
      startedAt:
        type: string
      targetDay:
        type: string
    type: object
  dangbamgong-backend_internal_dto.VoidHistoryResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.VoidSession'
        type: array
      targetDay:
        type: string
      totalDurationSec:
        type: integer
    type: object
  dangbamgong-backend_internal_dto.VoidSession:
    properties:
      activities:
        items:
          type: string
        type: array
      durationSec:
        type: integer
      endedAt:
        type: string
      sessionId:
        type: string
      startedAt:
        type: string
    type: object
...
  dangbamgong-backend_internal_dto.VoidStartResponse:
    properties:
      sessionId:
        type: string
      startedAt:
        type: string
      targetDay:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 밤의 공백 시간 측정 소셜 앱 API
  title: 당밤공 API
  version: "1.0"
paths:
...
  /users/{user_id}/block:
    post:
      description: 유저를 차단합니다. 기존 친구 관계 및 친구 요청이 모두 삭제됩니다.
      parameters:
      - description: 차단할 유저 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "409":
          description: ALREADY_BLOCKED
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 유저 차단
      tags:
      - Users
  /users/{user_id}/unblock:
    post:
      description: 차단을 해제합니다
      parameters:
      - description: 차단 해제할 유저 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: NOT_BLOCKED
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 유저 차단 해제
      tags:
      - Users
  /users/blocks:
    get:
      description: 내가 차단한 유저 목록을 반환합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_BlockListResponse'
      security:
      - BearerAuth: []
      summary: 차단 목록 조회
      tags:
      - Users
  /users/me:
    get:
      description: 현재 로그인한 유저의 프로필과 알림 설정을 반환합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UserMeResponse'
      security:
      - BearerAuth: []
      summary: 내 정보 조회
      tags:
      - Users
  /users/me/nickname:
    patch:
      consumes:
      - application/json
      description: 기존 닉네임을 변경합니다. 최초 설정은 /auth/nickname을 사용하세요.
      parameters:
      - description: 새 닉네임 (3-15자)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.ChangeNicknameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_ChangeNicknameResponse'
        "400":
          description: INVALID_NICKNAME
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 닉네임 변경
      tags:
      - Users
  /users/me/settings:
    patch:
      consumes:
      - application/json
      description: 알림 설정을 부분 업데이트합니다. 전달된 필드만 변경됩니다.
      parameters:
      - description: 변경할 설정
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UpdateSettingsResponse'
      security:
      - BearerAuth: []
      summary: 알림 설정 변경
      tags:
      - Users
  /users/search:
    get:
      description: 태그 접두사로 유저를 검색합니다. 나를 차단한 유저는 제외됩니다.
      parameters:
      - description: 태그 접두사
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UserSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 유저 검색
      tags:
      - Users
  /void/cancel:
    post:
      description: 진행 중인 공백을 취소합니다. 세션 기록이 남지 않습니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: NOT_IN_VOID
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 공백 취소
      tags:
      - Void
  /void/end:
    post:
      consumes:
      - application/json
      description: 진행 중인 공백 세션을 종료하고 활동을 기록합니다. 활동은 최대 5개.
      parameters:
      - description: 종료 시 기록할 활동 목록
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.VoidEndRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidEndResponse'
        "400":
          description: NOT_IN_VOID / TOO_MANY_ACTIVITIES / ACTIVITY_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 공백 종료
      tags:
      - Void
  /void/history:
    get:
      description: 특정 날짜의 공백 세션 목록과 총 시간을 반환합니다. 날짜 기준은 KST 16:00.
      parameters:
      - description: 조회할 날짜 (YYYY-MM-DD)
        in: query
        name: target_day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidHistoryResponse'
      security:
      - BearerAuth: []
      summary: 공백 히스토리 조회
      tags:
      - Void
  /void/start:
    post:
      description: 공백(밤의 공백) 세션을 시작합니다. 이미 공백 중이면 실패합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidStartResponse'
        "409":
          description: ALREADY_IN_VOID
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 공백 시작
      tags:
      - Void
  /void/test:
    post:
      consumes:
      - application/json
      description: 개발용. 임의의 시작/종료 시각으로 공백 세션을 생성합니다.
      parameters:
      - description: 테스트 공백 데이터
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.TestVoidRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_VoidEndResponse'
      security:
      - BearerAuth: []
      summary: 테스트 공백 생성
      tags:
      - Void
securityDefinitions:
  BearerAuth:
    description: Bearer {JWT 토큰}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"